Add ajax upvote-downvote, many minor adjustments to html and css

+++ app/controllers/answer.rb
@@ -16,20 +16,38 @@ get '/answers/:id/vote' do
   answer = Answer.find(params[:id])
   redirect "/questions/#{answer.question.id}" if !logged_in?
   if answer.votes.find_by(user_id: current_user.id) != nil
-    redirect "/questions/#{answer.question.id}"
+    if request.xhr?
+      answer.total_votes.to_s
+    else
+      redirect "/questions/#{answer.question.id}"
+    end
   else
     new_vote = Vote.new(voter: current_user)
     answer.votes << new_vote
+   if request.xhr?
+      answer.total_votes.to_s
+    else
+      redirect "/questions/#{answer.question.id}"
+    end
+  end
+
+  if request.xhr?
+    answer.total_votes.to_s
+  else
     redirect "/questions/#{answer.question.id}"
   end
 end
 
 get '/answers/:id/delete-vote' do
-  question = Answer.find(params[:id])
+  answer = Answer.find(params[:id])
   redirect "/answers/#{answer.id}" if !logged_in?
   vote = answer.votes.find_by(user_id: current_user.id)
-  vote.destroy
-  redirect "/questions/#{answer.question.id}"
+  vote.destroy if vote
+  if request.xhr?
+    answer.total_votes.to_s
+  else
+    redirect "/questions/#{answer.question.id}"
+  end
 end
 
 post '/answers/:id/comments' do
